memory/spi_mem: add spi memory api calls to read/write data
This adds an api that allows to easily read/write pages and data
buffers, erase sectors or the whole external flash memory, and verify if
the memory is erased.

+++ external/asf4-drivers/hpl/spi/spi_lite.h
@@ -89,17 +89,17 @@ uint32_t SPI_MEM_exchange_data(uint32_t data);
 /**
  * \brief Exchange block in SPI module
  */
-void SPI_MEM_exchange_block(void *block, uint8_t size);
+void SPI_MEM_exchange_block(void *block, size_t size);
 
 /**
  * \brief Write block in SPI module
  */
-void SPI_MEM_write_block(void *block, uint8_t size);
+void SPI_MEM_write_block(void *block, size_t size);
 
 /**
  * \brief Read block in SPI module
  */
-void SPI_MEM_read_block(void *block, uint8_t size);
+void SPI_MEM_read_block(void *block, size_t size);
 
 // Calculate baud register value from requested baudrate value
 #ifndef SERCOM3_BAUD_RATE
